Fix for rank calculation in extension parsing
Fixed a bug in riscv arch parsing where extensions starting with 'z'
were ordered in the wrong way. Exts starting with 'zi' and 'ze' were
ordered with the highest priority. This caused exts starting  with
'zb' to be written after those with 'zi'. The order of ext caused
problems when fowarding arch to asm.

Reworked the logic when calculating the ext rank in the ordering.